Integrate ecocredits module
